opea-comps/mega-service/requirements.txt
opea-comps
fastapi==0.68.1
uvicorn==0.15.0
requests==2.26.0
pydantic==1.8.2